requirements.txt
streamlit
streamlit-drawable-canvas
diffusers
transformers
torch
Pillow
opencv-python
ultralytics
numpy
xformers
